make # messages uint16_t removed unneeded code don't print text of deleted messages in wwivutil

+++ sdk/msgapi/message_area_wwiv.h
@@ -41,11 +41,11 @@ public:
   WWIVMessageAreaHeader(uint16_t wwiv_num_version, uint32_t active_message_count);
 
   const subfile_header_t& header() const { return header_; }
-  uint32_t active_message_count() const { return header_.active_message_count; }
-  void set_active_message_count(uint32_t active_message_count) {
+  uint16_t active_message_count() const { return header_.active_message_count; }
+  void set_active_message_count(uint16_t active_message_count) {
     header_.active_message_count = active_message_count;
   }
-  uint32_t increment_active_message_count() { return ++header_.active_message_count; }
+  uint16_t increment_active_message_count() { return ++header_.active_message_count; }
   bool initialized() const { return initialized_; }
   void set_initialized(bool initialized) { initialized_ = initialized; }
